Finalize initial implementation of comm module
Moved OptArgs into api struct (it didn't work otherwise).

Made a type for the api initialization functions and added a function
pointer to such a function on the module contexts. This way, the actual
program and the tests can define separate implementations of the api.

Moved relevant stuff from the old tests into new ones and copied the api
test implementations from the program.