Moved test code developed for this branch into src
Rationale:
Even though the design allows for the creation of a test api that runs
differently than the regular program, what I'm working on now *is* the
program, so it doesn't really make sense for it to be in tests. This
will also remove development friction since I don't have to update
certain things twice, once in src and once in tests.