Several updates to Sdb.h (see description)
Added macro that sets the log module instance pointer to NULL if logging
isn't going to be used. This is necessary since some Sdb.h macro
function wrappers pass the logging module instance pointer as an
argument.
Added NULL check for logging module pointer in the logging function. If
it is NULL, then the function returns.
Updated SdbLoadFileIntoMemory so it can use malloc if no arena is
provided.